requirements.txt
autopep8==1.7.0
pycodestyle==2.9.1
pysimdjson==5.0.1
toml==0.10.2
